+++ database/migrations/2023_08_22_070112_create_izin_kelola.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('no_SK');
             $table->unsignedBigInteger('id_ptk');
             $table->foreign('id_ptk')->references('id_ptk')->on('petak');
-            $table->decimal('luas_izin');
+            $table->double('luas_izin');
             $table->boolean('IsDelete')->default(0);
             $table->timestamps();
         });
